requirements.txt
ruff==0.14.1
pyzmq==27.1.0
click==8.3.0
flask==3.1.2
flask-socketio==5.4.1
python-socketio==5.11.4
httpx==0.27.2
